# docs/swagger.yaml
basePath: /api/v2
definitions:
...
  handler.BroadcastRequest:
    properties:
      message:
        type: string
    required:
    - message
    type: object
...
  handler.NotifySettingsRequest:
    properties:
      notify_expire:
        type: boolean
      notify_ticket:
        type: boolean
      notify_traffic:
        type: boolean
    type: object
  handler.SendNotificationRequest:
    properties:
      content:
        type: string
      message:
        type: string
      telegram_id:
        type: integer
      title:
        type: string
    type: object
  handler.SetWebhookRequest:
    properties:
      url:
        type: string
    type: object
...
  handler.UpdateBotRequest:
    properties:
      admin_ids: {}
      allow_bind:
        type: boolean
      allow_info:
        type: boolean
      allow_sub:
        type: boolean
      allow_ticket:
        type: boolean
      name:
        type: string
      token:
        type: string
      welcome_message:
        type: string
      welcome_msg:
        type: string
    type: object
...
  service.TelegramCallback:
    properties:
      data:
        type: string
      from:
        $ref: '#/definitions/service.TelegramUser'
      id:
        type: string
      message:
        $ref: '#/definitions/service.TelegramMessage'
    type: object
  service.TelegramChat:
    properties:
      id:
        type: integer
      title:
        type: string
      type:
        type: string
      username:
        type: string
    type: object
  service.TelegramMessage:
    properties:
      chat:
        $ref: '#/definitions/service.TelegramChat'
      date:
        type: integer
      from:
        $ref: '#/definitions/service.TelegramUser'
      message_id:
        type: integer
      text:
        type: string
    type: object
  service.TelegramUpdate:
    properties:
      callback_query:
        $ref: '#/definitions/service.TelegramCallback'
      message:
        $ref: '#/definitions/service.TelegramMessage'
      update_id:
        type: integer
    type: object
  service.TelegramUser:
    properties:
      first_name:
        type: string
      id:
        type: integer
      is_bot:
        type: boolean
      language_code:
        type: string
      last_name:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@anixops.com
    name: API Support
    url: https://github.com/AnixOps/anix-control/issues
  description: |-
    AnixOps Control 统一控制面 API 文档
    支持用户管理、节点管理、订阅系统、支付网关、流量转发等功能
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/AnixOps/anix-control
  title: AnixOps Control API
  version: 3.0.0-alpha.1
paths:
...
  /admin/telegram/bot:
    get:
      consumes:
      - application/json
      description: 管理员获取Telegram Bot的配置信�?
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取Telegram Bot配置
      tags:
      - 管理�?通知
    put:
      consumes:
      - application/json
      description: 管理员更新Telegram Bot的配置信�?
      parameters:
      - description: Bot配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateBotRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新Telegram Bot配置
      tags:
      - 管理�?通知
  /admin/telegram/broadcast:
    post:
      consumes:
      - application/json
      description: 管理员通过Telegram广播消息给所有绑定用�?
      parameters:
      - description: 广播请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.BroadcastRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 广播Telegram消息
      tags:
      - 管理�?通知
  /admin/telegram/notify:
    post:
      consumes:
      - application/json
      description: 管理员通过Telegram发送通知给指定用�?
      parameters:
      - description: 通知请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.SendNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 发送Telegram通知
      tags:
      - 管理�?通知
  /admin/telegram/users:
    get:
      consumes:
      - application/json
      description: 管理员获取Telegram用户绑定列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取用户绑定列表
      tags:
      - 管理�?通知
  /admin/telegram/users/{id}/notify:
    put:
      consumes:
      - application/json
      description: 绠＄悊鍛樻洿鏂版寚瀹歍elegram鐢ㄦ埛鐨勯€氱煡寮€鍏?// @Tags 绠＄悊绔?閫氱煡
      parameters:
      - description: Telegram缁戝畾ID
        in: path
        name: id
        required: true
        type: integer
      - description: 閫氱煡璁剧疆
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 鏇存柊Telegram鐢ㄦ埛閫氱煡璁剧疆
  /admin/telegram/webhook:
    delete:
      consumes:
      - application/json
      description: 管理员删除Telegram Bot的Webhook配置
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除Telegram Webhook
      tags:
      - 管理�?通知
    post:
      consumes:
      - application/json
      description: 管理员设置Telegram Bot的Webhook地址
      parameters:
      - description: Webhook配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.SetWebhookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 设置Telegram Webhook
      tags:
      - 管理�?通知
...
  /telegram/webhook:
    post:
      consumes:
      - application/json
      description: 接收Telegram Bot的Webhook回调
      parameters:
      - description: Telegram Update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.TelegramUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Telegram Webhook回调
      tags:
      - Telegram
...
  /user/telegram/notify:
    post:
      consumes:
      - application/json
      description: 用户更新自己的Telegram通知设置
      parameters:
      - description: 通知设置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.NotifySettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新Telegram通知设置
      tags:
      - 用户�?
  /user/telegram/status:
    get:
      consumes:
      - application/json
      description: 用户获取自己的Telegram绑定状�?
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取Telegram绑定状�?
      tags:
      - 用户�?
  /user/telegram/unbind:
    post:
      consumes:
      - application/json
      description: 用户解绑自己的Telegram账号
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 解绑Telegram
      tags:
      - 用户�?
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...